relocate serializer tests
