--- backend/api/tests/serializers/__init__.py
@@ -1,4 +1,0 @@
-"""
-© Ocado Group
-Created on 30/01/2024 at 19:03:37(+00:00).
-"""
